+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## (IN PROGRESS)
 
+* Results List | Hyperlink one column to improve accessibility. Refs UIREC-244.
+
 ## [2.3.0](https://github.com/folio-org/ui-receiving/tree/v2.3.0) (2022-10-27)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v2.2.0...v2.3.0)